Fix delete produto

+++ src/main/java/com/restaurante/grupo07/dto/produto/AddProdutoDto.java
@@ -14,7 +14,7 @@ public record AddProdutoDto(
         String foto,
 
         @NotNull
-        float valor,
+        double valor,
 
         @NotNull
         boolean disponivel,
